Backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
blis==0.7.10
cachetools==5.3.1
catalogue==1.0.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cymem==2.0.7
dataclasses-json==0.5.14
distlib==0.3.7
errant==2.3.3
faiss-cpu==1.7.4
filelock==3.12.3
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.4.0
fsspec==2023.6.0
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.32
gramformer @ git+https://github.com/PrithivirajDamodaran/Gramformer.git@23425cd2e98a919384cab6156af8adf1c9d0639a
greenlet==2.0.2
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
InstructorEmbedding==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain==0.0.270
langsmith==0.0.24
Levenshtein==0.21.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pipenv==2023.8.25
plac==1.1.3
platformdirs==3.10.0
pluggy==1.3.0
preshed==3.0.8
protobuf==4.24.1
pyarrow==12.0.1
pydantic==2.2.0
pydantic_core==2.6.0
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
PyPDF2==3.0.1
pyproject-api==1.6.1
pytesseract @ file:///D:/VSCode/Python%20%2B%20Angular/doc-analysis-app/Backend/pytesseract
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.21.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
rapidfuzz==3.2.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
spacy==2.3.9
SQLAlchemy==2.0.20
srsly==1.0.7
streamlit==1.25.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tesseract==0.1.3
thinc==7.4.6
threadpoolctl==3.2.0
tiktoken==0.4.0
timm==0.9.6
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tox==4.11.0
tqdm==4.66.1
transformers==4.32.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.21.2
virtualenv==20.24.3
wasabi==0.10.1
watchdog==3.0.0
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
